some changes in ui_form, complite ui_server_settings_window_form.py
